Implement the action for clicking effect.
