+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/Toolbar/ToolbarButton.cs
@@ -1,15 +1,12 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
+using System.Diagnostics.CodeAnalysis;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Input;
-using osu.Framework.Input.Events;
-using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osuTK;
 
@@ -24,13 +21,10 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar
             IconContainer.Show();
         }
 
-        [Resolved]
+        [Resolved, AllowNull]
         private TextureStore textures { get; set; }
 
-        [Resolved]
-        private OsuColour colours { get; set; }
-
-        [Resolved]
+        [Resolved, AllowNull]
         private ReadableKeyCombinationProvider keyCombinationProvider { get; set; }
 
         public void SetIcon(string texture) =>
@@ -44,18 +38,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar
             {
                 Icon = iconUsage
             });
-
-        protected void ToggleClickEffect()
-        {
-            if (Enabled.Value)
-            {
-                IconContainer.FadeOut(100).Then().FadeIn();
-            }
-            else
-            {
-                IconContainer.FadeColour(colours.Red, 100).Then().FadeColour(Colour4.White);
-            }
-        }
 
         protected void SetState(bool enabled)
         {
@@ -77,14 +59,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar
                     Alpha = 0,
                 },
             };
-        }
-
-        protected override bool OnClick(ClickEvent e)
-        {
-            if (Enabled.Value)
-                ToggleClickEffect();
-
-            return base.OnClick(e);
         }
     }
 }
